ecommerce/requirements/base.txt
amqp==2.6.1
asgiref==3.2.10
billiard==3.6.3.0
celery==4.4.5
certifi==2020.6.20
chardet==3.0.4
defusedxml==0.6.0
Django==3.0
django-allauth==0.42.0
django-ckeditor==6.0.0
django-etc==1.2.0
django-hitcount==1.3.3
django-js-asset==1.2.2
djangorestframework==3.11.1
freeze==3.0
future==0.18.2
idna==2.10
importlib-metadata==2.0.0
kombu==4.6.11
mysqlclient==2.0.1
oauthlib==3.1.0
Pillow==7.2.0
python3-openid==3.2.0
pytz==2020.1
PyYAML==5.3.1
redis==3.5.3
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.10
vine==1.3.0
zipp==3.2.0
